requirements.txt
altgraph==0.16.1
future==0.16.0
macholib==1.10
mysql-connector-python==8.0.12
pefile==2018.8.8
protobuf==3.6.1
PyInstaller==3.3.1
pywin32==223
six==1.11.0
